+++ src/git.rs
@@ -5,6 +5,7 @@ use tempfile::NamedTempFile;
 use crate::{
     credential::{Credential, CredentialType},
     job::JobResult,
+    log::LogLevel,
     utils::{execute_command, execute_command_with_env},
 };
 
@@ -16,12 +17,15 @@ pub fn git_clone(
     job_result: &mut JobResult,
 ) -> Result<(), String> {
     if cfg!(target_os = "windows") {
-        // Workaround for local
-        execute_command(
-            &format!("git clone -b {} {}", branch, url),
-            directory.clone(),
-            job_result,
-        )?;
+        if !job_result.dry_run {
+            // Workaround for local
+            execute_command(
+                &format!("git clone -b {} {}", branch, url),
+                directory.clone(),
+                job_result,
+            )?;
+        }
+
         Ok(())
     } else if credential_id.is_some() {
         let credential = Credential::get(credential_id.unwrap(), Some(job_result))?;
@@ -30,26 +34,30 @@ pub fn git_clone(
         }
         let credential = credential.unwrap();
         if let CredentialType::Ssh(ssh_credential) = credential.value {
-            let tmp_file = NamedTempFile::new().map_err(|e| e.to_string())?;
-            let tmp_path = tmp_file.path();
-            let _ = std::fs::write(tmp_path, ssh_credential.private_key);
+            job_result.add_log(LogLevel::Info, format!("command: chmod 400 <private_key_temp_file>"));
+            job_result.add_log(LogLevel::Info, format!("command: git clone -b {} {}", branch, url));
+            if !job_result.dry_run {
+                let tmp_file = NamedTempFile::new().map_err(|e| e.to_string())?;
+                let tmp_path = tmp_file.path();
+                let _ = std::fs::write(tmp_path, ssh_credential.private_key);
 
-            execute_command(
-                &format!("chmod 400 {}", tmp_path.display()),
-                directory.clone(),
-                job_result,
-            )?;
+                execute_command(
+                    &format!("chmod 400 {}", tmp_path.display()),
+                    directory.clone(),
+                    job_result,
+                )?;
 
-            let env = vec![(
-                "GIT_SSH_COMMAND".to_string(),
-                format!("ssh -i {} -o StrictHostKeyChecking=no", tmp_path.display()),
-            )];
-            execute_command_with_env(
-                &format!("git clone -b {} {}", branch, url),
-                directory.clone(),
-                env,
-                job_result,
-            )?;
+                let env = vec![(
+                    "GIT_SSH_COMMAND".to_string(),
+                    format!("ssh -i {} -o StrictHostKeyChecking=no", tmp_path.display()),
+                )];
+                execute_command_with_env(
+                    &format!("git clone -b {} {}", branch, url),
+                    directory.clone(),
+                    env,
+                    job_result,
+                )?;
+            }
 
             Ok(())
         } else {
